green tests corrected

+++ src/main/java/org/app/Calculator.java
@@ -1,19 +1,19 @@
 package org.app;
 
 public class Calculator {
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
 
-    public int sub(int a, int b) {
+    public double sub(double a, double b) {
         return a - b;
     }
 
-    public int mul(int a, int b) {
+    public double mul(double a, double b) {
         return a * b;
     }
 
-    public int div(int a, int b) throws ArithmeticException {
+    public double div(double a, double b) throws ArithmeticException {
         return a / b;
     }
 }
